+++ cmd/cortextool/pkg/commands/rules.go
@@ -58,6 +58,8 @@ type RuleCommand struct {
 	RuleFilesPath string
 
 	// Sync/Diff Rules Config
+	Namespaces           string
+	namespacesMap        map[string]struct{}
 	IgnoredNamespaces    string
 	ignoredNamespacesMap map[string]struct{}
 
@@ -157,7 +159,9 @@ func (r *RuleCommand) Register(app *kingpin.Application) {
 	loadRulesCmd.Arg("rule-files", "The rule files to check.").Required().ExistingFilesVar(&r.RuleFilesList)
 
 	// Diff Command
-	diffRulesCmd.Flag("ignored-namespaces", "comma-separated list of namespaces to ignore during a diff.").StringVar(&r.IgnoredNamespaces)
+	diffRulesCmd.Arg("rule-files", "The rule files to check.").ExistingFilesVar(&r.RuleFilesList)
+	diffRulesCmd.Flag("namespaces", "comma-separated list of namespaces to check during a diff. Cannot be used together with --ignored-namespaces.").StringVar(&r.Namespaces)
+	diffRulesCmd.Flag("ignored-namespaces", "comma-separated list of namespaces to ignore during a diff. Cannot be used together with --namespaces.").StringVar(&r.IgnoredNamespaces)
 	diffRulesCmd.Flag("rule-files", "The rule files to check. Flag can be reused to load multiple files.").StringVar(&r.RuleFiles)
 	diffRulesCmd.Flag(
 		"rule-dirs",
@@ -166,7 +170,9 @@ func (r *RuleCommand) Register(app *kingpin.Application) {
 	diffRulesCmd.Flag("disable-color", "disable colored output").BoolVar(&r.DisableColor)
 
 	// Sync Command
-	syncRulesCmd.Flag("ignored-namespaces", "comma-separated list of namespaces to ignore during a sync.").StringVar(&r.IgnoredNamespaces)
+	syncRulesCmd.Arg("rule-files", "The rule files to check.").ExistingFilesVar(&r.RuleFilesList)
+	syncRulesCmd.Flag("namespaces", "comma-separated list of namespaces to check during a diff. Cannot be used together with --ignored-namespaces.").StringVar(&r.Namespaces)
+	syncRulesCmd.Flag("ignored-namespaces", "comma-separated list of namespaces to ignore during a sync. Cannot be used together with --namespaces.").StringVar(&r.IgnoredNamespaces)
 	syncRulesCmd.Flag("rule-files", "The rule files to check. Flag can be reused to load multiple files.").StringVar(&r.RuleFiles)
 	syncRulesCmd.Flag(
 		"rule-dirs",
@@ -221,11 +227,27 @@ func (r *RuleCommand) setup(k *kingpin.ParseContext) error {
 }
 
 func (r *RuleCommand) setupFiles() error {
+	if r.Namespaces != "" && r.IgnoredNamespaces != "" {
+		return errors.New("--namespaces and --ignored-namespaces cannot be set at the same time")
+	}
+
 	// Set up ignored namespaces map for sync/diff command
-	r.ignoredNamespacesMap = map[string]struct{}{}
-	for _, ns := range strings.Split(r.IgnoredNamespaces, ",") {
-		if ns != "" {
-			r.ignoredNamespacesMap[ns] = struct{}{}
+	if r.IgnoredNamespaces != "" {
+		r.ignoredNamespacesMap = map[string]struct{}{}
+		for _, ns := range strings.Split(r.IgnoredNamespaces, ",") {
+			if ns != "" {
+				r.ignoredNamespacesMap[ns] = struct{}{}
+			}
+		}
+	}
+
+	// Set up allowed namespaces map for sync/diff command
+	if r.Namespaces != "" {
+		r.namespacesMap = map[string]struct{}{}
+		for _, ns := range strings.Split(r.Namespaces, ",") {
+			if ns != "" {
+				r.namespacesMap[ns] = struct{}{}
+			}
 		}
 	}
 
@@ -372,6 +394,18 @@ func (r *RuleCommand) loadRules(k *kingpin.ParseContext) error {
 	return nil
 }
 
+// shouldCheckNamespace returns whether the namespace should be checked according to the allowed and ignored namespaces
+func (r *RuleCommand) shouldCheckNamespace(namespace string) bool {
+	// when we have an allow list, only check those that we have explicitly defined.
+	if r.namespacesMap != nil {
+		_, allowed := r.namespacesMap[namespace]
+		return allowed
+	}
+
+	_, ignored := r.ignoredNamespacesMap[namespace]
+	return !ignored
+}
+
 func (r *RuleCommand) diffRules(k *kingpin.ParseContext) error {
 	err := r.setupFiles()
 	if err != nil {
@@ -391,6 +425,10 @@ func (r *RuleCommand) diffRules(k *kingpin.ParseContext) error {
 	changes := []rules.NamespaceChange{}
 
 	for _, ns := range nss {
+		if !r.shouldCheckNamespace(ns.Namespace) {
+			continue
+		}
+
 		currentNamespace, exists := currentNamespaceMap[ns.Namespace]
 		if !exists {
 			changes = append(changes, rules.NamespaceChange{
@@ -413,13 +451,15 @@ func (r *RuleCommand) diffRules(k *kingpin.ParseContext) error {
 	}
 
 	for ns, deletedGroups := range currentNamespaceMap {
-		if _, ignored := r.ignoredNamespacesMap[ns]; !ignored {
-			changes = append(changes, rules.NamespaceChange{
-				State:         rules.Deleted,
-				Namespace:     ns,
-				GroupsDeleted: deletedGroups,
-			})
+		if !r.shouldCheckNamespace(ns) {
+			continue
 		}
+
+		changes = append(changes, rules.NamespaceChange{
+			State:         rules.Deleted,
+			Namespace:     ns,
+			GroupsDeleted: deletedGroups,
+		})
 	}
 
 	p := printer.New(r.DisableColor)
@@ -445,6 +485,10 @@ func (r *RuleCommand) syncRules(k *kingpin.ParseContext) error {
 	changes := []rules.NamespaceChange{}
 
 	for _, ns := range nss {
+		if !r.shouldCheckNamespace(ns.Namespace) {
+			continue
+		}
+
 		currentNamespace, exists := currentNamespaceMap[ns.Namespace]
 		if !exists {
 			changes = append(changes, rules.NamespaceChange{
@@ -467,13 +511,15 @@ func (r *RuleCommand) syncRules(k *kingpin.ParseContext) error {
 	}
 
 	for ns, deletedGroups := range currentNamespaceMap {
-		if _, ignored := r.ignoredNamespacesMap[ns]; !ignored {
-			changes = append(changes, rules.NamespaceChange{
-				State:         rules.Deleted,
-				Namespace:     ns,
-				GroupsDeleted: deletedGroups,
-			})
+		if !r.shouldCheckNamespace(ns) {
+			continue
 		}
+
+		changes = append(changes, rules.NamespaceChange{
+			State:         rules.Deleted,
+			Namespace:     ns,
+			GroupsDeleted: deletedGroups,
+		})
 	}
 
 	err = r.executeChanges(context.Background(), changes)
@@ -488,6 +534,10 @@ func (r *RuleCommand) executeChanges(ctx context.Context, changes []rules.Namesp
 	var err error
 	for _, ch := range changes {
 		for _, g := range ch.GroupsCreated {
+			if !r.shouldCheckNamespace(ch.Namespace) {
+				continue
+			}
+
 			log.WithFields(log.Fields{
 				"group":     g.Name,
 				"namespace": ch.Namespace,
@@ -499,6 +549,10 @@ func (r *RuleCommand) executeChanges(ctx context.Context, changes []rules.Namesp
 		}
 
 		for _, g := range ch.GroupsUpdated {
+			if !r.shouldCheckNamespace(ch.Namespace) {
+				continue
+			}
+
 			log.WithFields(log.Fields{
 				"group":     g.New.Name,
 				"namespace": ch.Namespace,
@@ -510,6 +564,10 @@ func (r *RuleCommand) executeChanges(ctx context.Context, changes []rules.Namesp
 		}
 
 		for _, g := range ch.GroupsDeleted {
+			if !r.shouldCheckNamespace(ch.Namespace) {
+				continue
+			}
+
 			log.WithFields(log.Fields{
 				"group":     g.Name,
 				"namespace": ch.Namespace,