+++ cmd/cortextool/CHANGELOG.md
@@ -4,6 +4,8 @@
 ## Unreleased
 
 * [BUGFIX] Fix inaccuracy in `e2ealerting` caused by invalid purging condition on timestamps. #117
+* [FEATURE] Add support for position rule-files arguments to `rules sync` and `rules diff` #125
+* [FEATURE] Add an allow-list of namespaces for `rules sync` and `rules diff` #125
 
 ## v0.5.0